Fix Baro read
Add missing cast to uint16_t. This is needed because on 32bit platforms default result of subtraction is 32 bit, but the comparison reguires 16 bit for it to be overflow proof. Without that reads will not work after about the minute from the startup.

+++ speeduino/sensors.ino
@@ -245,7 +245,7 @@ while(adcState !=ADCrunning) //do not leave the scene until we have gotten the A
       }
       break;
     case BaroAdc:
-      if((currentMillis - lastmillisBARO) >= BARO_INTERVAL) //Infrequent baro readings are not an issue.
+      if((uint16_t)(currentMillis - lastmillisBARO) >= BARO_INTERVAL) //Infrequent baro readings are not an issue.
       {
         adcState = readBaro(adcState);        
         if (adcState == ADCidle)
